requirements.txt
--index-url https://download.pytorch.org/whl/cu118

# python 3.10.0
torch==2.0.0+cu118
transformers==4.28.1
hydra-core==1.3.2
mlflow==2.2.2
scikit-learn==1.2.2
datasets==2.11.0
evaluate==0.4.0